Eliminate remaining double-precision math across firmware
Comprehensive pass to remove software-emulated double arithmetic on
the ESP32-S3 (single-precision FPU only).

IMU330 (208 Hz, biggest single win):
  ACCEL_RES/GYRO_RES #defines lacked f suffixes and parentheses —
  6 double multiplications every IMU cycle (1,248/sec eliminated).

EKF6 predict()+correct() (208 Hz):
  std::sin/cos/fabs → sinf/cosf/fabsf.  8 trig + 3 fabs calls per
  cycle were dispatching to software double routines.

KalmanFilter (208 Hz):
  std::fabs → fabsf; stray double literal.

OnSpeedTypes.h (208 Hz via safeAsin):
  std::asin/atan2/sqrt → asinf/atan2f/sqrtf.

AHRS.cpp EKF6 branch (208 Hz):
  sin/cos → sinf/cosf; double literal 0.0 → 0.0f throughout.

SensorIO (50 Hz):
  sqrt() → sqrtf(), double literals → float.

Audio (50 Hz + init):
  Remaining double literals in SetPulseFreq, SetVolume, constructor.

DisplaySerial (10 Hz):
  String compare g_Config.sSerialOutFormat → cached enum
  enSerialOutFormat (same pattern as bCalSourceEfis);
  floor() → floorf().

gLimit (10 Hz): fabs → fabsf, 0.666 → 2.0f/3.0f.
3DAudio (10 Hz): macro float + parens fix, abs → fabsf, literals.
MadgwickFusion: DEG2RAD macro float, cos/sin → cosf/sinf (init-only).
CurveCalc: std::log/exp → logf/expf (rare code paths).
BoomSerial: String(val,HEX) → %02X format; constructor literals.

Flash code shrinks ~4.7 KB beyond the initial commit on this branch.
All 114 native tests pass.

+++ software/Libraries/onspeed_core/CurveCalc.cpp
@@ -38,7 +38,7 @@ float CurveCalc(float x, const SuCalibrationCurve& curve) {
     else if (curve.iCurveType == 2) {
         if (x <= 0.0f)
             return 0.0f;
-        y = curve.afCoeff[MAX_CURVE_COEFF - 2] * std::log(x)
+        y = curve.afCoeff[MAX_CURVE_COEFF - 2] * logf(x)
           + curve.afCoeff[MAX_CURVE_COEFF - 1];
         ONSPEED_LOG_DEBUG("%.2f * log(%.2f) + %.2f = %.2f\n",
             curve.afCoeff[MAX_CURVE_COEFF - 2], x,
@@ -48,7 +48,7 @@ float CurveCalc(float x, const SuCalibrationCurve& curve) {
     // Uses last two coefficients: afCoeff[2] = a, afCoeff[3] = b
     else if (curve.iCurveType == 3) {
         y = curve.afCoeff[MAX_CURVE_COEFF - 2]
-          * std::exp(curve.afCoeff[MAX_CURVE_COEFF - 1] * x);
+          * expf(curve.afCoeff[MAX_CURVE_COEFF - 1] * x);
         ONSPEED_LOG_DEBUG("%.2f * exp(%.2f * %.2f) = %.2f\n",
             curve.afCoeff[MAX_CURVE_COEFF - 2],
             curve.afCoeff[MAX_CURVE_COEFF - 1], x, y);
